+++ frontend/src/Settings/ImportLists/ImportListExclusions/ImportListExclusions.css.d.ts
@@ -3,9 +3,11 @@
 interface CssExports {
   'addButton': string;
   'addImportListExclusion': string;
+  'actions': string;
   'foreignId': string;
   'importListExclusionsHeader': string;
   'name': string;
+  'select': string;
 }
 export const cssExports: CssExports;
 export default cssExports;