+++ frontend/src/Settings/ImportLists/ImportListExclusions/ImportListExclusions.js
@@ -1,11 +1,15 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
+import Button from 'Components/Link/Button';
+import ConfirmModal from 'Components/Modal/ConfirmModal';
+import CheckInput from 'Components/Form/CheckInput';
 import FieldSet from 'Components/FieldSet';
 import Icon from 'Components/Icon';
 import Link from 'Components/Link/Link';
 import PageSectionContent from 'Components/Page/PageSectionContent';
-import { icons } from 'Helpers/Props';
+import { icons, kinds, sizes } from 'Helpers/Props';
 import translate from 'Utilities/String/translate';
+import getSelectedIds from 'Utilities/Table/getSelectedIds';
 import EditImportListExclusionModalConnector from './EditImportListExclusionModalConnector';
 import ImportListExclusion from './ImportListExclusion';
 import styles from './ImportListExclusions.css';
@@ -19,7 +23,11 @@ class ImportListExclusions extends Component {
     super(props, context);
 
     this.state = {
-      isAddImportListExclusionModalOpen: false
+      isAddImportListExclusionModalOpen: false,
+      isSelecting: false,
+      isBulkDeleteModalOpen: false,
+      allSelected: false,
+      selectedState: {}
     };
   }
 
@@ -34,6 +42,89 @@ class ImportListExclusions extends Component {
     this.setState({ isAddImportListExclusionModalOpen: false });
   };
 
+  onToggleSelectPress = () => {
+    const { items } = this.props;
+    const selectedState = {};
+
+    items.forEach((item) => {
+      selectedState[item.id] = false;
+    });
+
+    this.setState({
+      isSelecting: true,
+      selectedState,
+      allSelected: false
+    });
+  };
+
+  onDoneSelectingPress = () => {
+    this.setState({
+      isSelecting: false,
+      isBulkDeleteModalOpen: false,
+      selectedState: {},
+      allSelected: false
+    });
+  };
+
+  onSelectedChange = ({ id, value }) => {
+    this.setState((prevState) => {
+      const selectedState = {
+        ...prevState.selectedState,
+        [id]: value
+      };
+
+      const selectedIds = getSelectedIds(selectedState);
+      const allSelected = selectedIds.length > 0 && selectedIds.length === Object.keys(selectedState).length;
+
+      return {
+        selectedState,
+        allSelected
+      };
+    });
+  };
+
+  onSelectAllChange = ({ value }) => {
+    this.setState((prevState) => {
+      const selectedState = {};
+      Object.keys(prevState.selectedState).forEach((id) => {
+        selectedState[id] = value;
+      });
+
+      return {
+        selectedState,
+        allSelected: value
+      };
+    });
+  };
+
+  onDeleteSelectedPress = () => {
+    const selectedIds = getSelectedIds(this.state.selectedState);
+    if (!selectedIds.length) {
+      return;
+    }
+
+    this.setState({ isBulkDeleteModalOpen: true });
+  };
+
+  onBulkDeleteConfirm = () => {
+    const selectedIds = getSelectedIds(this.state.selectedState);
+    if (!selectedIds.length) {
+      this.setState({ isBulkDeleteModalOpen: false });
+      return;
+    }
+
+    this.props.deleteImportListExclusions(selectedIds);
+    this.setState({
+      isBulkDeleteModalOpen: false,
+      selectedState: {},
+      allSelected: false
+    });
+  };
+
+  onBulkDeleteCancel = () => {
+    this.setState({ isBulkDeleteModalOpen: false });
+  };
+
   //
   // Render
 
@@ -41,8 +132,18 @@ class ImportListExclusions extends Component {
     const {
       items,
       onConfirmDeleteImportListExclusion,
+      isDeleting,
       ...otherProps
     } = this.props;
+
+    const {
+      isSelecting,
+      isBulkDeleteModalOpen,
+      allSelected,
+      selectedState
+    } = this.state;
+
+    const selectedIds = getSelectedIds(selectedState);
 
     return (
       <FieldSet legend={translate('ImportListExclusions')}>
@@ -50,7 +151,48 @@ class ImportListExclusions extends Component {
           errorMessage={translate('UnableToLoadImportListExclusions')}
           {...otherProps}
         >
+          <div className={styles.actions}>
+            {!isSelecting && (
+              <Button
+                kind={kinds.PRIMARY}
+                size={sizes.SMALL}
+                onPress={this.onToggleSelectPress}
+              >
+                {translate('ExclusionSelect')}
+              </Button>
+            )}
+
+            {isSelecting && (
+              <>
+                <Button
+                  kind={kinds.DANGER}
+                  size={sizes.SMALL}
+                  isDisabled={!selectedIds.length || isDeleting}
+                  onPress={this.onDeleteSelectedPress}
+                >
+                  {translate('DeleteSelected')}
+                </Button>
+                <Button
+                  kind={kinds.DEFAULT}
+                  size={sizes.SMALL}
+                  onPress={this.onDoneSelectingPress}
+                >
+                  {translate('DoneSelecting')}
+                </Button>
+              </>
+            )}
+          </div>
+
           <div className={styles.importListExclusionsHeader}>
+            {isSelecting && (
+              <div className={styles.select}>
+                <CheckInput
+                  name="selectAllImportListExclusions"
+                  value={allSelected}
+                  onChange={this.onSelectAllChange}
+                />
+              </div>
+            )}
             <div className={styles.foreignId}>
               {translate('ForeignId')}
             </div>
@@ -69,6 +211,9 @@ class ImportListExclusions extends Component {
                     {...otherProps}
                     index={index}
                     onConfirmDeleteImportListExclusion={onConfirmDeleteImportListExclusion}
+                    isSelecting={isSelecting}
+                    isSelected={selectedState[item.id]}
+                    onSelectedChange={this.onSelectedChange}
                   />
                 );
               })
@@ -89,6 +234,16 @@ class ImportListExclusions extends Component {
             onModalClose={this.onModalClose}
           />
 
+          <ConfirmModal
+            isOpen={isBulkDeleteModalOpen}
+            kind={kinds.DANGER}
+            title={translate('DeleteImportListExclusion')}
+            message={translate('DeleteImportListExclusionMessageText')}
+            confirmLabel={translate('Delete')}
+            onConfirm={this.onBulkDeleteConfirm}
+            onCancel={this.onBulkDeleteCancel}
+          />
+
         </PageSectionContent>
       </FieldSet>
     );
@@ -99,7 +254,9 @@ ImportListExclusions.propTypes = {
   isFetching: PropTypes.bool.isRequired,
   error: PropTypes.object,
   items: PropTypes.arrayOf(PropTypes.object).isRequired,
-  onConfirmDeleteImportListExclusion: PropTypes.func.isRequired
+  onConfirmDeleteImportListExclusion: PropTypes.func.isRequired,
+  deleteImportListExclusions: PropTypes.func.isRequired,
+  isDeleting: PropTypes.bool.isRequired
 };
 
 export default ImportListExclusions;