+++ src/Readarr.Api.V1/ImportLists/ImportListExclusionBulkResource.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace Readarr.Api.V1.ImportLists
+{
+    public class ImportListExclusionBulkResource
+    {
+        public List<int> Ids { get; set; }
+    }
+}